# Conflicts:
#	package.json
#	src/modules/App/components/__snapshots__/App.test.js.snap
#
src/modules/SharedComponents/ContributorsEditor/components/ContributorFo
rm.js
#
src/modules/SharedComponents/ContributorsEditor/components/ContributorFo
rm.snapshot.test.js
#
src/modules/SharedComponents/ContributorsEditor/components/ContributorRo
w.js
#
src/modules/SharedComponents/ContributorsEditor/components/ContributorRo
w.snapshot.test.js
#
src/modules/SharedComponents/ContributorsEditor/components/ContributorRo
wHeader.snapshot.test.js
#
src/modules/SharedComponents/ContributorsEditor/components/ContributorsE
ditor.test.js
#
src/modules/SharedComponents/ContributorsEditor/components/__snapshots__
/ContributorForm.snapshot.test.js.snap
#
src/modules/SharedComponents/ContributorsEditor/components/__snapshots__
/ContributorRow.snapshot.test.js.snap
#
src/modules/SharedComponents/ContributorsEditor/components/__snapshots__
/ContributorRowHeader.snapshot.test.js.snap
#
src/modules/SharedComponents/ContributorsEditor/components/__snapshots__
/ContributorsEditor.test.js.snap
#	src/modules/SharedComponents/LookupFields/containers/UqIdField.js
#
src/modules/SharedComponents/Toolbox/ListEditor/components/FreeTextForm.
js
#
src/modules/SharedComponents/Toolbox/ListEditor/components/ListEditor.js
#
src/modules/SharedComponents/Toolbox/ListEditor/components/__snapshots__
/FreeTextForm.test.js.snap

+++ src/modules/SharedComponents/ContributorsEditor/components/ContributorRowHeader.js
@@ -16,6 +16,7 @@ export class ContributorRowHeader extends PureComponent {
     static propTypes = {
         onDeleteAll: PropTypes.func.isRequired,
         showIdentifierLookup: PropTypes.bool,
+        showRoleInput: PropTypes.bool,
         showContributorAssignment: PropTypes.bool,
         locale: PropTypes.object,
         disabled: PropTypes.bool,
@@ -28,6 +29,7 @@ export class ContributorRowHeader extends PureComponent {
             contributorAssignmentColumn: 'Select your name',
             nameColumn: 'Name as published',
             identifierColumn: 'UQ identifier',
+            roleColumn: 'Role',
             reorderColumn: 'Reorder records',
             deleteAll: 'Remove all records',
             deleteAllConfirmation: {
@@ -49,8 +51,8 @@ export class ContributorRowHeader extends PureComponent {
     };
 
     render() {
-        const {nameColumn, identifierColumn, reorderColumn, deleteAll, deleteAllConfirmation} = this.props.locale;
-        const {classes, showIdentifierLookup, isInfinite} = this.props;
+        const {nameColumn, identifierColumn, reorderColumn, deleteAll, deleteAllConfirmation, roleColumn} = this.props.locale;
+        const {classes, showIdentifierLookup, isInfinite, showRoleInput} = this.props;
         return (
             <Fragment>
                 <ConfirmDialogBox
@@ -77,6 +79,12 @@ export class ContributorRowHeader extends PureComponent {
                         <Hidden xsDown>
                             <ListItemText secondary={'Affiliation %'} secondaryTypographyProps={{variant: 'caption'}}/>
                             <ListItemText secondary={identifierColumn} secondaryTypographyProps={{variant: 'caption'}}/>
+                        </Hidden>
+                    }
+                    {
+                        showRoleInput &&
+                        <Hidden xsDown>
+                            <ListItemText secondary={roleColumn} secondaryTypographyProps={{variant: 'caption'}}/>
                         </Hidden>
                     }
                     <Hidden xsDown>
